fix question

+++ src/main/java/com/myproject/repository/QuestionRepository.java
@@ -20,6 +20,6 @@ public interface QuestionRepository {
     Question getQuestionById(int questionId);
     List<Object> getListQuestionsForQuestion(Map<String, String> params);
 //    List<Object> getListQuestionsForQuestionAndAnswer();
-    int countQuetionsNotLive();
+    Long countQuetionsNotLive();
     List<Question> getQuestions(Map<String, String> params);
 }
